Module foundry_config::error

error handling and solc error codes

Structs

ExtractConfigError

Represents a failed attempt to extract Config from a Figment

Enums

FoundryConfigError

Represents an error that can occur when constructing the Config

SolidityErrorCode

A non-exhaustive list of solidity error codes

Constants

FAILED_TO_EXTRACT_CONFIG_PANIC_MSG

The message shown upon panic if the config could not be extracted from the figment