Module foundry_config::etherscan

Support for multiple Etherscan keys.

Structs

EtherscanConfig

Represents all info required to create an etherscan client

EtherscanConfigs

Container type for Etherscan API keys and URLs.

EtherscanEnvProvider 🔒

A [Provider] that provides Etherscan API key from the environment if it’s not empty.

ResolvedEtherscanConfig

Contains required url + api key to set up an etherscan client

ResolvedEtherscanConfigs

Container type for resolved etherscan keys, see EtherscanConfigs::resolved.

Enums

EtherscanApiKey

Represents a single etherscan API key

EtherscanConfigError

Errors that can occur when creating an EtherscanConfig

Constants

ETHERSCAN_USER_AGENT

The user agent to use when querying the etherscan API.

Functions

into_url 🔒

This is a hack to work around IntoUrl’s sealed private functions, which can’t be called normally.