Module foundry_config::filter

Helpers for constructing and using [FileFilter]s.

Structs

GlobMatcher

A globset::Glob that creates its globset::GlobMatcher when its created, so it doesn’t need to be compiled when the filter functions TestFilter functions are called.

SkipBuildFilters

Bundles multiple SkipBuildFilter into a single FileFilter

Enums

SkipBuildFilter

A filter that excludes matching contracts from the build

Functions

expand_globs

Expand globs with a root path.