Struct foundry_config::fix::TomlFile

struct TomlFile {
    doc: DocumentMut,
    path: PathBuf,
}

A convenience wrapper around a TOML document and the path it was read from

Fields
doc: DocumentMutpath: PathBuf
Implementations

impl TomlFile

fn open(path: impl AsRef<Path>) -> Result<Self>

fn doc(&self) -> &DocumentMut

fn doc_mut(&mut self) -> &mut DocumentMut

fn path(&self) -> &Path

fn save(&self) -> Result<()>

impl TomlFile

fn insert_profile( &mut self, profile_str: &str, value: Item, ) -> Result<(), InsertProfileError>

Insert a name as [profile.name]. Creating the [profile] table where necessary and throwing an error if there exists a conflict