Module foundry_config::inline

Re-exports

pub use conf_parser::parse_config_bool;

pub use conf_parser::parse_config_u32;

pub use conf_parser::validate_profiles;

pub use conf_parser::InlineConfigParser;

pub use error::InlineConfigError;

pub use error::InlineConfigParserError;

pub use natspec::NatSpec;

Modules

conf_parser 🔒

error 🔒

natspec 🔒

Structs

InlineConfig

Represents per-test configurations, declared inline as structured comments in Solidity test files. This allows to create configs directly bound to a solidity test.

Constants

INLINE_CONFIG_FUZZ_KEY

INLINE_CONFIG_INVARIANT_KEY

INLINE_CONFIG_PREFIX 🔒

Statics

INLINE_CONFIG_PREFIX_SELECTED_PROFILE 🔒

Functions

remove_whitespaces 🔒