foundry_config::inline
Module conf_parser

Traits

InlineConfigParser

This trait is intended to parse configurations from structured text. Foundry users can annotate Solidity test functions, providing special configs just for the execution of a specific test.

Functions

parse_config_bool

Tries to parse a bool from value. The key argument is used to give details in the case of an error.

parse_config_u32

Tries to parse a u32 from value. The key argument is used to give details in the case of an error.

validate_profiles

Checks if all configuration lines specified in natspec use a valid profile.