foundry_config

Macro merge_impl_figment_convert

macro_rules! merge_impl_figment_convert {
    ($name:ty, $start:ident $(, $more:ident)*) => { ... };
}

Same as impl_figment_convert but also merges the type itself into the figment

Example

Merge several nested Provider together with the type itself

use foundry_config::{
    Config,
    figment::{value::*, *},
    impl_figment_convert, merge_impl_figment_convert,
};
use std::path::PathBuf;

#[derive(Default)]
struct MyArgs {
    root: Option<PathBuf>,
}

impl Provider for MyArgs {
    fn metadata(&self) -> Metadata {
        Metadata::default()
    }

    fn data(&self) -> std::result::Result<Map<Profile, Dict>, Error> {
        todo!()
    }
}

impl_figment_convert!(MyArgs);

#[derive(Default)]
struct OuterArgs {
    value: u64,
    inner: MyArgs,
}

impl Provider for OuterArgs {
    fn metadata(&self) -> Metadata {
        Metadata::default()
    }

    fn data(&self) -> std::result::Result<Map<Profile, Dict>, Error> {
        todo!()
    }
}

merge_impl_figment_convert!(OuterArgs, inner);