foundry_config::providers

Module ext

Structs

BackwardsCompatTomlProvider 🔒

A Provider that handles breaking changes in toml files

DappEnvCompatProvider 🔒

A provider that checks for DAPP_ env vars that are named differently than FOUNDRY_

DappHardhatDirProvider 🔒

A provider that sets the src and output path depending on their existence.

FallbackProfileProvider

Extracts the profile from the profile key and sets unset values according to the fallback provider

ForcedSnakeCaseData 🔒

A Provider that ensures all keys are snake case if they’re not standalone sections, See Config::STANDALONE_SECTIONS

OptionalStrictProfileProvider 🔒

Extracts the profile from the profile key and using the original key as backup, merging values where necessary

RenameProfileProvider 🔒

Renames a profile from from to to.

TomlFileProvider 🔒

A convenience provider to retrieve a toml file. This will return an error if the env var is set but the file does not exist

UnwrapProfileProvider 🔒

Unwraps a profile reducing the key depth

WrapProfileProvider 🔒

Wraps a profile in another profile

Traits

ProviderExt 🔒