foundry_config
Module soldeer

Configuration specific to the forge soldeer command and the forge_soldeer package

Structs

MapDependency

Soldeer dependencies config structure when it’s defined as a map

SoldeerConfig

Type for Soldeer configs, under soldeer tag in the foundry.toml

SoldeerDependencyConfig

Type for Soldeer configs, under dependencies tag in the foundry.toml

Enums

RemappingsLocation

Location where to store the remappings, either in remappings.txt or in the foundry.toml

SoldeerDependencyValue

Enum to cover both available formats for defining a dependency dep = { version = "1.1", url = "https://my-dependency" } or dep = "1.1"

Functions

default_true 🔒