foundry_config

Struct Chain

pub struct Chain(ChainKind);

Either a known NamedChain or a EIP-155 chain ID.

Tuple Fields
0: ChainKind
Implementations

impl Chain

pub const fn from_named(named: NamedChain) -> Chain

Creates a new Chain by wrapping a NamedChain.

pub fn from_id(id: u64) -> Chain

Creates a new Chain by wrapping a NamedChain.

pub const fn is_named(self) -> bool

Returns true if this a named variant.

pub const fn is_id(self) -> bool

Returns true if this an Id variant.

pub const fn from_id_unchecked(id: u64) -> Chain

Creates a new Chain from the given ID, without checking if an associated NamedChain exists.

This is discouraged, as other methods assume that the chain ID is not known, but it is not unsafe.

pub const fn mainnet() -> Chain

Returns the mainnet chain.

pub const fn goerli() -> Chain

Returns the goerli chain.

pub const fn holesky() -> Chain

Returns the holesky chain.

pub const fn hoodi() -> Chain

Returns the hoodi chain.

pub const fn sepolia() -> Chain

Returns the sepolia chain.

pub const fn optimism_mainnet() -> Chain

Returns the optimism mainnet chain.

pub const fn optimism_goerli() -> Chain

Returns the optimism goerli chain.

pub const fn optimism_sepolia() -> Chain

Returns the optimism sepolia chain.

pub const fn base_mainnet() -> Chain

Returns the base mainnet chain.

pub const fn base_goerli() -> Chain

Returns the base goerli chain.

pub const fn base_sepolia() -> Chain

Returns the base sepolia chain.

pub const fn arbitrum_mainnet() -> Chain

Returns the arbitrum mainnet chain.

pub const fn arbitrum_nova() -> Chain

Returns the arbitrum nova chain.

pub const fn arbitrum_goerli() -> Chain

Returns the arbitrum goerli chain.

pub const fn arbitrum_sepolia() -> Chain

Returns the arbitrum sepolia chain.

pub const fn arbitrum_testnet() -> Chain

Returns the arbitrum testnet chain.

pub const fn syndr() -> Chain

Returns the syndr l3 mainnet chain.

pub const fn syndr_sepolia() -> Chain

Returns the syndr sepolia chain.

pub const fn fraxtal() -> Chain

Returns the fraxtal mainnet chain.

pub const fn fraxtal_testnet() -> Chain

Returns the fraxtal testnet chain.

pub const fn blast() -> Chain

Returns the blast chain.

pub const fn blast_sepolia() -> Chain

Returns the blast sepolia chain.

pub const fn linea() -> Chain

Returns the linea mainnet chain.

pub const fn linea_goerli() -> Chain

Returns the linea goerli chain.

pub const fn linea_sepolia() -> Chain

Returns the linea sepolia chain.

pub const fn mode() -> Chain

Returns the mode mainnet chain.

pub const fn mode_sepolia() -> Chain

Returns the mode sepolia chain.

pub const fn elastos() -> Chain

Returns the elastos mainnet chain.

pub const fn degen() -> Chain

Returns the degen l3 mainnet chain.

pub const fn dev() -> Chain

Returns the dev chain.

pub const fn bsc_mainnet() -> Chain

Returns the bsc mainnet chain.

pub const fn bsc_testnet() -> Chain

Returns the bsc testnet chain.

pub const fn opbnb_mainnet() -> Chain

Returns the opbnb mainnet chain.

pub const fn opbnb_testnet() -> Chain

Returns the opbnb testnet chain.

pub const fn ronin() -> Chain

Returns the ronin mainnet chain.

pub const fn ronin_testnet() -> Chain

Returns the ronin testnet chain.

pub const fn taiko() -> Chain

Returns the taiko mainnet chain.

pub const fn taiko_hekla() -> Chain

Returns the taiko hekla chain.

pub const fn shimmer() -> Chain

Returns the shimmer testnet chain.

pub const fn flare() -> Chain

Returns the flare mainnet chain.

pub const fn flare_coston2() -> Chain

Returns the flare testnet chain.

pub const fn darwinia() -> Chain

Returns the darwinia mainnet chain.

pub const fn crab() -> Chain

Returns the crab mainnet chain.

pub const fn koi() -> Chain

Returns the koi testnet chain.

pub const fn immutable() -> Chain

Returns the Immutable zkEVM mainnet chain.

pub const fn immutable_testnet() -> Chain

Returns the Immutable zkEVM testnet chain.

pub const fn ink_sepolia() -> Chain

Returns the ink sepolia chain.

pub const fn ink_mainnet() -> Chain

Returns the ink mainnet chain.

pub const fn scroll_mainnet() -> Chain

Returns the scroll mainnet chain

pub const fn scroll_sepolia() -> Chain

Returns the scroll sepolia chain

pub const fn treasure() -> Chain

Returns the Treasure mainnet chain.

pub const fn treasure_topaz_testnet() -> Chain

Returns the Treasure Topaz testnet chain.

pub const fn berachain() -> Chain

Returns the Berachain mainnet chain.

pub const fn berachain_bepolia() -> Chain

Returns the Berachain Bepolia testnet chain.

pub const fn sonic() -> Chain

Returns the Sonic mainnet chain.

pub const fn sonic_blaze() -> Chain

Returns the Sonic Blaze testnet chain.

pub const fn superposition_testnet() -> Chain

Returns the Superposition testnet chain.

pub const fn superposition() -> Chain

Returns the Superposition mainnet chain.

pub const fn unichain_mainnet() -> Chain

Returns the Unichain mainnet chain.

pub const fn unichain_sepolia() -> Chain

Returns the Unichain sepolia chain.

pub const fn zksync() -> Chain

Returns the ZKSync mainnet chain.

pub const fn zksync_testnet() -> Chain

Returns the ZKSync testnet chain.

pub const fn abs() -> Chain

Returns the Abstract mainnet chain.

pub const fn abstract_testnet() -> Chain

Returns the Abstract testnet chain.

pub const fn sophon() -> Chain

Returns the Sophon mainnet chain.

pub const fn sophon_testnet() -> Chain

Returns the Sophon testnet chain.

pub const fn lens() -> Chain

Returns the Lens mainnet chain.

pub const fn lens_testnet() -> Chain

Returns the Lens testnet chain.

pub const fn kind(&self) -> &ChainKind

Returns the kind of this chain.

pub const fn into_kind(self) -> ChainKind

Returns the kind of this chain.

pub const fn is_ethereum(&self) -> bool

Returns true if this chain is Ethereum or an Ethereum testnet.

pub const fn is_optimism(self) -> bool

Returns true if the chain contains Optimism configuration.

pub const fn is_polygon(&self) -> bool

Returns true if this chain is a Polygon chain.

pub const fn is_arbitrum(self) -> bool

Returns true if the chain contains Arbitrum configuration.

pub const fn is_elastic(self) -> bool

Returns true if the chain contains Elastic Network configuration.

pub const fn named(self) -> Option<NamedChain>

Attempts to convert the chain into a named chain.

pub const fn id(self) -> u64

The ID of the chain.

impl Chain

Methods delegated to NamedChain. Note that [ChainKind::Id] won’t be converted because it was already done at construction.

pub const fn average_blocktime_hint(self) -> Option<Duration>

Returns the chain’s average blocktime, if applicable.

See NamedChain::average_blocktime_hint for more info.

pub const fn is_legacy(self) -> bool

Returns whether the chain implements EIP-1559 (with the type 2 EIP-2718 transaction type).

See NamedChain::is_legacy for more info.

pub const fn supports_shanghai(self) -> bool

Returns whether the chain supports the Shanghai hardfork.

See NamedChain::supports_shanghai for more info.

pub const fn etherscan_urls(self) -> Option<(&'static str, &'static str)>

Returns the chain’s blockchain explorer and its API (Etherscan and Etherscan-like) URLs.

See NamedChain::etherscan_urls for more info.

pub const fn etherscan_api_key_name(self) -> Option<&'static str>

Returns the chain’s blockchain explorer’s API key environment variable’s default name.

See NamedChain::etherscan_api_key_name for more info.

pub fn etherscan_api_key(self) -> Option<String>

Returns the chain’s blockchain explorer’s API key, from the environment variable with the name specified in etherscan_api_key_name.

See NamedChain::etherscan_api_key for more info.

pub fn public_dns_network_protocol(self) -> Option<String>

Returns the address of the public DNS node list for the given chain.

See NamedChain::public_dns_network_protocol for more info.