Struct foundry_config::ResolvedRpcEndpoints

pub struct ResolvedRpcEndpoints {
    endpoints: BTreeMap<String, Result<String, UnresolvedEnvVarError>>,
}

Container type for resolved endpoints, see RpcEndpoint::resolve.

Fields
endpoints: BTreeMap<String, Result<String, UnresolvedEnvVarError>>

contains all named endpoints and their URL or an error if we failed to resolve the env var alias

Implementations

impl ResolvedRpcEndpoints

pub fn has_unresolved(&self) -> bool

Returns true if theres an endpoint that couldnt be resolved