Module foundry_config::utils

Utility functions

Enums

Numeric

Helper type to parse both u64 and U256

Functions

deserialize_stringified_percent 🔒

Deserialize stringified percent. The value must be between 0 and 100 inclusive.

deserialize_usize_or_max 🔒

Deserialize an usize or

evm_spec_id

Returns the [SpecId] derived from [EvmVersion]

find_git_root_path

Returns the path of the top-level directory of the working git tree. If there is no working tree, an error is returned.

find_project_root_path

Returns the root path to set for the project root

foundry_toml_dirs

Returns a list of unique paths to all folders under root that contain a foundry.toml file

get_available_profiles

Returns all available profile keys in a given .toml file

get_dir_remapping 🔒

Returns a remapping for the given dir

load_config

Loads the config for the current project workspace

load_config_with_root

Loads the config for the current project workspace or the provided root path

read_toml 🔒

Returns a [toml_edit::Document] loaded from the provided path. Can raise an error in case of I/O or parsing errors.

remappings_from_env_var

Returns the remappings from the given var

remappings_from_newline

Returns all [Remapping]s contained in the remappings str separated by newlines

to_array_value

Converts the val into a figment::Value::Array