foundry_config::utils
Function find_project_root

pub fn find_project_root(cwd: Option<&Path>) -> PathBuf

Returns the root path to set for the project root.

Traverse the dir tree up and look for a foundry.toml file starting at the given path or cwd, but only until the root dir of the current repo so that:

-- foundry.toml

-- repo
  |__ .git
  |__sub
     |__ [given_path | cwd]

will still detect repo as root.

Returns repo or cwd if no foundry.toml is found in the tree.

Panics

Panics if:

cwd is Some and is not a valid directory;

cwd is None and the std::env::current_dir call fails.