foundry_config

Module inline

Re-exports

pub use natspec::*;

Modules

natspec 🔒

Structs

InlineConfig

Represents per-test configurations, declared inline as structured comments in Solidity test files. This allows to create configs directly bound to a solidity test.

InlineConfigError

Wrapper error struct that catches config parsing errors, enriching them with context information reporting the misconfigured line.

InlineConfigProvider

[figment::Provider] for InlineConfig at a given contract and function level.

Enums

InlineConfigErrorKind

Errors returned when parsing inline config.

Constants

INLINE_CONFIG_PREFIX 🔒

Functions

extend_data_map 🔒

extend_dict 🔒

extend_value 🔒

Type Aliases

DataMap 🔒